backend/requirements.txt
fastapi
uvicorn[standard]
openai>=1.42.0
python-dotenv
pydantic>=2
gunicorn
